+++ Test/error_test.py
@@ -65,11 +65,11 @@ pycm.pycm_obj.pycmCompareError: The domain of all ConfusionMatrix objects must b
 >>> cp = Compare({"cm1":cm,"cm2":cm3},by_class=True,weight={1:1,2:1})
 Traceback (most recent call last):
         ...
-pycm.pycm_obj.pycmCompareError: The weight type must be dictionary and also must be set for all classes.
+pycm.pycm_obj.pycmCompareError: The weight type must be dictionary and also must be specified for all of the classes.
 >>> cp = Compare({"cm1":cm,"cm2":cm3},by_class=True,weight=[])
 Traceback (most recent call last):
         ...
-pycm.pycm_obj.pycmCompareError: The weight type must be dictionary and also must be set for all classes.
+pycm.pycm_obj.pycmCompareError: The weight type must be dictionary and also must be specified for all of the classes.
 >>> cm.CI("MCC")
 Traceback (most recent call last):
         ...
@@ -79,4 +79,16 @@ Supported parameters : TPR,TNR,PPV,NPV,ACC,PLR,NLR,FPR,FNR,AUC,PRE,Kappa,Overall
 Traceback (most recent call last):
         ...
 pycm.pycm_obj.pycmCIError: The input type is considered to be string but it's not!
+>>> cm.weighted_average("AUCC")
+Traceback (most recent call last):
+        ...
+pycm.pycm_error.pycmAverageError: Invalid parameter!
+>>> cm.weighted_average("AUC",weight=2)
+Traceback (most recent call last):
+        ...
+pycm.pycm_error.pycmAverageError: The weight type must be dictionary and also must be specified for all of the classes.
+>>> cm.weighted_average("AUC",weight={1:23})
+Traceback (most recent call last):
+        ...
+pycm.pycm_error.pycmAverageError: The weight type must be dictionary and also must be specified for all of the classes.
 """