+++ Test/function_test.py
@@ -495,6 +495,10 @@ True
 0.4
 >>> IS_calc(13,0,0,38)
 1.5474877953024933
+>>> weighted_kappa_calc(cm2.classes,cm2.table,cm2.P,cm2.TOP,cm2.POP,cm2.table)
+-0.3883495145631068
+>>> weighted_kappa_calc(cm2.classes,cm2.table,cm2.P,cm2.TOP,cm2.POP,{1:{1:2,2:2}})
+'None'
 >>> kappa_no_prevalence_calc(cm2.Overall_ACC)
 0.33333333333333326
 >>> reliability_calc(cm2.Overall_RACC,cm2.Overall_ACC)