+++ pycm/pycm_param.py
@@ -39,6 +39,8 @@ COMPARE_WEIGHT_ERROR = "The weight type must be dictionary and also must be spec
 
 COMPARE_RESULT_WARNING = "Confusion matrices are too close and the best one can not be recognized."
 
+WEIGHTED_KAPPA_WARNING = "The weight format is wrong, the result is for unweighted kappa."
+
 AVERAGE_WEIGHT_ERROR = "The weight type must be dictionary and also must be specified for all of the classes."
 
 AVERAGE_INVALID_ERROR = "Invalid parameter!"