+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Document modified
 - `relabel` method bug fixed
+- `matrix_check` function bug fixed
 - Minor bug in `Compare` class fixed
 - Document modified
 ## [2.6] - 2020-03-25