+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Github action
 ### Changed
 - Document modified
+- `README.md` modified
 - `relabel` method bug fixed
 - `matrix_check` function bug fixed
 - Minor bug in `Compare` class fixed