+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `weighted_average` method
 - `pycmAverageError` class
+- MATLAB examples
+- Github action
 ### Changed
 - Document modified
+- `README.md` modified
 - `relabel` method bug fixed
+- `matrix_check` function bug fixed
 - Minor bug in `Compare` class fixed
 ## [2.6] - 2020-03-25
 ### Added