+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `weighted_average` method
+- `weighted_kappa` method
 - `pycmAverageError` class
 - MATLAB examples
 - Github action