+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Document modified
 - `README.md` modified
 - `relabel` method bug fixed
+- `sparse_table_print` function bug fixed
 - `matrix_check` function bug fixed
 - Minor bug in `Compare` class fixed
 - Class names mismatch bug fixed